Added new HBase configurations and Bug Fix 1391962
New HBase configuration settings added.

Bug fix 1391962: A check has been added that makes sure that Trafodion
is installed on the same nodes that the Hadoop distro is installed on.
If they do not match trafodion_mods will exit and ask user to reinstall
using a correct list of nodes.

Change-Id: I0d5a2bca7636f3e7b3c4baa66c93b7e6c9c2d27b