Change HSTableDef to recognize volatile tables
The redefinition of the isVolatile() virtual function for
some subclasses of HSTableDef originally returned FALSE, and
had not been updated since the addition of volatile tables.
Now the function always determines whether the table is volatile
based on the name of the containing schema.

The symptom of the bug was a core file produced for certain
operations on a virtual table, such as bulk load.

Change-Id: I1d592935d4922e8ce8202bf7429f929275ebae0f
Closes-Bug: #1388207