Change default for USTAT_USE_IS_WHEN_NO_STATS
This cqd was originally set to OFF by default, causing
internal sort not to be used in cases where there are
no existing histograms. Internal sort is known to
underperform when there are a small number of unique
values for a column, and without prior histograms, the
unique entry count is unknown.

However, experience with POCs has shown that using internal
sort in these situations (unique count unknown) is in
general significantly faster than not using it, so the
default setting has been changed to ON.

Running developer regressions with this setting exposed a
bug that occurs when internal sort is used on an empty table
that previously did contain rows. This change includes a known
diffs file for the failing test until the bug (LP 1393930)
can be fixed.

Change-Id: I9ff6672a168016fb83fadb26b40a30c7fa3eee6c
Closes-Bug: #1392555