Query Invalidation triggered by DDL, phase 3
This check-in allows invalidation of queries which have been prepared
and are held by the EXE for  execution or re-execution. It does not
invalidate running queries.

When an attempt is made to (re)execute an invalidated query, a
special SQLCODE, 8738, is raised and the query is sent ack to the
compiler via the AQR mechanism. The check-in include a new test cases
in the executor/TEST122 regression test which demonstrate the
functionality.

Change-Id: I98ea484f9d7fc9e43372e3057d8e8288162865c1
Implements: blueprint ddl-query-invalidation