SQL syntax to cancel executing query, phase 1
This check-in implements the command "control query cancel qid <qid>".
The command will send a cancel message to the MXSSMP cancel broker,
but since the master executor is not yet sending a query started
message,
no query will be canceled, and if the message is successful SQLCODE
-8026 will be raised. The change also validates the cancel command.
produces new EXPLAIN and SHOWPLAN output and handles errors, as shown
in a new regression test.

Change-Id: I472fca043aa12d684c13581264baf22a7ea5f49b
Implements: blueprint sql-query-cancel