Fix performance regression due to QI for DDL
This check-in omits object UIDs from query plans for the tables
SB_HISTOGRAMS and SB_HISTOGRAMS_INTERVALS. Previously, when the
code generator tried to add the object UIDs for these, it had to
make a special query to the metadata, since the corresponding
internal cached structure omitted object UIDs when they were
created via methods like Generator::createVirtualTableDesc. The
special query to lookup these object UIDs was shown to be
responsible for a large pathlengh regression.

Change-Id: Id5046c5c55a4fc8dd2ba3f891449ea87d35a5534
Closes-Bug: #1398600