Updated Phoenix tests to cope with ANSI schemas
Updated Phoenix tests to do 'create schema' and 'drop schema' before and after
each test to cope with the new behavior with ANSI schemas.  Up to now, schemas
in Trafodion do not need to be created or dropped, but the ANSI schema support
will change that.   This update for Phoenix tests need to go in first before
the implement of ANSI schemas can go in, to ensure a smooth transition for
Jenkins builds.  Since 'create schema' and 'drop schema' are just no-op right
now, Phoenix tests will continue to work for the existing builds even with this
update.

Change-Id: Ifa6902a122c78bdb899516cf8121295092cad0a9