Miscellaneous minor changes
1) Revert a part of previous change to ComTdbHbaseAccess::displayRowId() for
SHOWPLAN. Each valid row id should have 2 bytes of length followed by the
value of id, if it is not missing or null. We need to treat the id as
binary value so we'll print each id byte as 2-digit hex value up to 20
bytes.

2) Remove the part of comments in sql/sqlcomp/DefineConstants.h that
requires not to delete and reuse existing constants. This is no longer
mandatory in Trafodion.

3) Enable NA memory overflow checking when running executor and qat
suites in debug flavor. This would increase the run time for these two
suites by about 10%.

Rework according to review comments. Explanations above changed too.

Change-Id: I854ec776fa5fb815c39677d469f8f46f18d51402