Fix install_local_hadoop to find maven, run phoenix T2
Use the MAVEN variable set in sqenvcom.sh, rather than relying on it being
in the PATH.

Change exit code of DCS failure, since it is not required to run
trafodion instance.

Add in the option to run phoenix tests with jdbc T2 driver, now that
Alice Chen has enhanced the test script to work in local_hadoop environment.

Change-Id: I3ebc5fe194d1c28b98a777ad276fbdcd036f857a