Automated collection of necessary statistics
When the optimizer requests a histogram for a given column, and that
histogram does not exist, it may (depending on the cqds in effect) register
a request for the histogram to be created at a later time, or utilize a
small sample to generate a rudimentary histogram on the fly. In either
case, when a subsequent Update Statistics statements specifies the ON
NECESSARY COLUMNS clause, any column of the target table that has been
the subject of one of these actions will have a bona fide histogram
created.

blueprint ustat-automation

Change-Id: Ieceac3e1d84bf8091a5bf340dc1739a447ad2436