SQL syntax to cancel executing query, phase 2
This check-in causes queries to register with the cancel
broker using a nowait Seabed message, and to process the
completion of the message by canceling the executing query
if the cancel broker so indicates in the message reply. The
check-in includes new test cases in the regression test to
cancel different types of SELECT queries.

These types of queries do not yet respond to cancel:
1. Queries that are decomposed into child queries, e.g.,
CREATE INDEX, CREATE TABLE AS, UPDATE STATISTICS, etc;
2. UPDATE and DELETE
3. INSERT queries might work now, but haven't been tested.

By design, unique queries cannot be canceled.

Implements: blueprint sql-query-cancel

    Change-Id: I95aaf9b42de645b60ba68cdce0da729b84f1eb82

Change-Id: I1ee1f05fd33c2cfb72405a5ddddb9518ab52076d