Manageability changes in DCS
1) Port mapping
At startup DcsMaster now creates a list of TCP/IP ports based on the
entries in the conf/servers file and writes it to the zookeeper
/<username>/trafodion/dcs/servers/registered znode. Each MXOSRVR reads
the port map list and retrieves its unique port number.

2) Trafodion repository
JDBC type4 is used to access SQL from DcsMaster. The feature is disabled
by default but may be enabled by overriding the unpublished
"dcs.master.trafodion.repository" property with a value of .true. in
your dcs-site.xml file.

3) HA for DcsMaster
A new file, /bin/master-backup.sh, was added that when called from
dcs-start/dcs-stop script look for the presence of optional
/conf/backup-masters file. This feature works like the existing
.servers. in that you can specify hostnames, one per line, e.g.,
.localhost.. The script will start/stop DcsMaster processes on all hosts
defined in the file. The DcsMaster has new leader election code so that
only one DcsMaster process is the .leader.. All others wait, in a
non-herding fashion, for other DcsMasters to fail. The vision is that
the Seaquest .ndcsbind. script may be run  prior to starting DCS but
more effort is needed to test this. This commit sets up the
infrastructure. There.s no property for this feature as it.s enabled
only the /conf/backup-masters file exists and has host names defined.

4) Trafodion logs
As a convenience DcsMaster can display the Trafodion log files found in
$MY_SQROOT/logs. There was no requirement for this rather I added for my
development use. Be warned that DcsMaster is not aware of Trafodion
configuration changes so right now it only looks in $MY_SQROOT/logs.
This feature is disabled by default but may be enabled by overriding the
unpublished "dcs.master.trafodion.logs" property with a value of .true.
in your dcs-site.xml file.

5) Basic Security
Provide basic infrastructure to authenticate DcsMaster Web UI users. A
new file /conf/realm.properties was added as a first step. It allows for
definition of users with roles. The embedded Jetty web server reads this
file and uses it to authenticate users attempting to access the
DcsMaster web UI. This feature will surely change over time. It also
supports #3 above as we need username/password for the connection to SQL
in T4 driver. We want users to only see/cancel their own queries. This
feature is disabled by default but may be enabled by overriding the
unpublished "dcs.master.security" property with a value of .true. in
your dcs-site.xml file.

Change-Id: I3e258f698dd828e683f0de03c2fe67c3ce4a5d26