Ustat fails with duplicate CLI statement name
Under certain circumstances, processing of the ON NECESSARY
COLUMNS clause will occur while another dynamic cursor is
still open, causing the default statement name used by
HSCursor to be duplicated. When instantiating HSCursor for
NECESSARY, a unique statement name is now supplied to the
constructor.

Another unrelated change is included; the access clause for
the query to retrieve necessary columns now specifies FOR
COMMITTED ACCESS rather than FOR UNCOMMITTED ACCESS. This
is a response to a code review comment for the delivery of
ustat automation that was not included in that delivery
because of Git problems.

Change-Id: I574b62d0d665fd4b6fd79fd0c6184050bef82d63