Pass object type to NATableDB:removeNATable
This change fixes a bug which can happen when indexes or
sequence generators are dropped and, due to NATable::objectUID_
not being initialize, the NATableDB:removeNATable method must
lookup the object UID from metadata. Before this change, the
code assumed anything looked up was a base table. Now the
object type is passed to NATableDB:removeNATable.

Change-Id: I4566ee68aa84367ea33c2e2bb1d3d6a346e8ce97
Closes-Bug: #1401999