Fix LP bug 1389347 - mxosrvr occasionally dropping core.
QA found that the mxosrvr process occasionally drops core ...
usually when they ran the Phoenix tests, but sometimes when
they ran other QA tests.

The problem was that mxosrvr was not being linked with libjsig.
Consequently, JAVA's signal handling for SIGSEGV (and maybe others)
was conflicting with LLVM's signal handling for the same signals.
The fix is to link libjsig into mxosrvr very early on in the
linking process ... just as we did for the SQL Compiler.

File modified: conn/odbc/src/odbc/nsksrvr/Makefile

Fix was pre-reviewed by Arvind.

Change-Id: I64df21db5d3f8292b0634480dfc31a257efd2856