Fix for LP bug 1404951
This is an intermittent problem that appears on the build machine,
caused by Change-Id I5b570c42712d4c38157181c3b76bf9a3ab6e2ed9. In this
delivery we are increasing the number of rows fetched by hbase scan call
to two rows, up from the previous value of one row. This is only for the
scan call used to determine mdam probe keys.

Change-Id: I12f20084bf53c188000db24ed8a698b3fdc7f41b