Fix for bug 1408187 missing commas in SQL event log files
The QRLogger::log method did not provide values for SQLCode and Query ID
that the event log reader TMUDF expects. In some cases, the calls to
QRLogger::log provided these values in the text, or they provided
two commas to indicate NULL values, but there are hundreds of other
calls to QRLogger::log not doing either of these. Added a second
QRLogger::log method and handled SQLCode and QueryID logging inside
those methods, so only a few calls had to be changed.

Change-Id: I9d147398df3056ed5c8b04be20ceebd3f42fe006