Fix for bug 1409128
Reworked the code to delete Compiler Contexts (CmpContext) from
the Trafodion SQL Engine CLI contexts (ContextCli). Now all the
CmpContexts created in a ContextCli are deallocated when it is deleted
as part of JDBC disconnect call.

Earlier, the code was deallocating  the current CmpContext alone. However
the current CmpContext may not be from the current ContextCli. This was
causing the deallocated memory to be referenced again.

Change-Id: Ia957f17a8733e528495107b097d29bdc92d7bf95