Fix for bug 1355032
Problem is that SPJ result set virtual table could not resolve column
names when same column names exists in two different tables.  The
problem is in binder as SPJ result set virtual table columns names were
not renamed in table desc. Because of this when two tables were joined
in an SPJ that have the same column names and could not be resolved.
Renamed SPJ virtual table column names to C1, C2, C3.....

Change-Id: Iabd1e8d2e4aadaca344ee47c07b66e38b641b996