Fix for bug 1404108 where mxosrvrs do not exit on a sqstop
When sqstop or dcsstop is called some mxosrvrs do not exit when the
statistics collection feature is enabled. The statistics collection
occurs in a separate thread and that thread sometimes does not
terminate and also holds on to a mutex and does not get released during
exit. The fix is to retry the thread termination during exit time.

Fixes bug 1404108

Change-Id: Ia5764a4e3672da8dbade00f8e4b5bdacbb0d5ac5