Reworked fix for LP bug 1404951
The scan cache size for an mdam probe is now set to the hbase default of 100.
Setting it values like 1 or 2 resulted in intermittent failures. The cqd
COMP_BOOL_184 can be set ON to get a cache size of 1 for mdam probes.
Root cause for this intermittent failure will be investigated later.

Change-Id: Ic05a77ecb0deeb260784f156de251a0f0dbdf49c