skipping seabase/test025 test for R1.0.
build/test seem to be hanging off and on after running
this test or during this test.
we dont know what the root cause is, it only happens
on build machines.
disabling this test for now. Cant say for sure if that
will take care of hang but we will try that out.
This test is for IDENTITY cols which is a tech preview
feature.
Post 1.0, we will enable this test and find/fix the root
cause.

Change-Id: I0dc541c74a7a8a27fd55e6dc14b9f839f6f9d8a3