Fix for LP bugs 1404442 & 1401683
These are duplicate problems.

A QA test suite (arkcase/arkt1112.log)  contains some update statistics
tests which are starting a transaction but the transaction is never
ended.

Since the rogue transaction is started in a secondary arkcmp, the master
arkcmp is unaware of this transaction.

What differentiates this from other tests is that QA test turns off
autocommit  before performing a series of update statistics.

The fix is to encapsulate a begin/commit work around DDL operations
executed by update statistics.

Change-Id: Ibbed2133e5a79f7393c9ec1b25c884a7dd8b0c3e