Fix for bug #1410928. Mxosrvr coring during performance tests.
The core seems to be occurring because of a lack of synchronization
between the timer thread and main thread accessing a global statement
object. When the statement is dropped the associated pSrvrStmt
is also deleted and hence the global pointer is invalidated and causes
the core when accessed. The fix now also nulls the global statement
pointer so that the timer thread ignores the dropped
statement.

Fixes bug #1410928

Change-Id: I06b15b90325a7b405d4adcca871b58c9dba51729