Fix bug 1413422: Error 8605 for stmt QRYSTS_ON_1 during connect.
mxosrvr does an explicit COMMIT WORK after fetching the setting
for a mode since during this time AUTOCOMMIT is off and SHOWCONTROL
used to start a transaction.

Recent  change disabled starting a transaction for SHOWCONTROL
leading to the 8605 error.
(in GenRelScan.cpp, method Describe::codeGen)
https://review.trafodion.org/938.

Do note this fix is only for the following seen in master logs:

SQLCODE: 8605, QID: xxx_STMT_QRYSTS_ON_1, *** ERROR[8605]
Committing a transaction which has not started.

Change-Id: Idb047b55d575f49b77b7660ab7716b32de48edf0