Fix for NATable Heap leak
Earlier, each NATable entry had its own heap. Now, there is a NATable Heap
created for each compiler context. All the cached NATable object is
created from this heap.

This is done to reduce the memory usage in the compiler context and to
take care of leak of NATable heap.

Change-Id: I3ec362c0b8e5b1466956c3773c2b5dc7dccf0d52