Merge the private shared single script repo into master.
This is a checkin of the single script project which basically
wraps all the various scripts into one single script that the user
runs.  The script now has the logic to determine which Hadoop distro
is installed as well as to determine if it is a full install or only
an upgrade and many other changes.

This was developed and tested in a private shared repo so
a review of the code has already been done.  QA has approved this
to be checked in.

edit1: fixed shift problem in parsing parameters
edit2: Fixed problems with user config file
       Added more error checking for config file
edit3: Fixed config directory creation
edit4: Fixed copyright and OpenStack VM package installation
edit5: Changed Java version detection
       No longer do yum removes if package is there already
edit6: Refixed OpenStack VM package installation that got
          overlaid in a merge
       Added chown command to handle case where SQ_ROOT was not
          a sub-directory of the Trafodion userid's home dir

Change-Id: I37a331c89c311feee430541b69f5f7dfe2177e60