Change that possibly fixes the bug 1418757
If a scanner is closed before the next call completes, it
results in OutOfOrderScannerNextException. Trafodion engine can
close the scanner when the next is active in a different thread
with pre-fetch concept. The fix is to cancel the pending next
before the scanner is closed.

Thanks to Oliver and Joanie for tracking this down. Ran
catman1/TEST137 many times successfully

Change-Id: Ifb080abdd9bbe4042c88e2df1fa07b17f8eba2b5