Change that possibly fixes the bug 1418757
If a scanner is closed before the next call completes, it
results in OutOfOrderScannerNextException. Trafodion engine can
close the scanner when the next is active in a different thread
with pre-fetch concept. The fix is to cancel the pending next
before the scanner is closed.

Thanks to Oliver and Joanie for tracking this down. Ran
catman1/TEST137 many times successfully

Change-Id: I3b2872e0ca4fc6586e7e486600748c1102994e7b