Fix to block txn on split/close region bug 1274651
The closing variable needed to be implemented since the
move from subclassing.  This adds it back to the
preSplit() and preClose() calls to block new transactions
that try to begin on the region.
This closes a timing window where the split
or close operation starts and a transactions begins and
fails with various exceptions, now it will have the
exception message:
closing region, no more new transactions allowed.
Region: <regionName>

Change-Id: Id0dbe8c28c5fc69050e2fca184c77dcea9097f83