Fix for bug #1415936
Bug #1415936:host name is blank for mxosrvr INFO log.
To fix it ,call host_addr_get to get host name no matter
whether -TCPADD is set or not.
If TCPADD is set and different with the ip returned
by host_addr_get,TCPADD is used and host name is set to "UNRESOLVED".
If TCPADD is set and same with the ip returned by host_addr_get,
TCPADD is used and host name is set from host_addr_get.
If TCPADD is not set ,ip and host name are set from host_addr_get.

Change-Id: I987dd156e6fc16f99c1d2b24e17d6d041cd3bcb1