Full backup and restore utilities
Shell scripts to perform full offline backup and restore
of all trafodion tables including meta data tables. Trafodion
needs to be shutdown before doing the backup or restore operations.
The Backup operation takes snapshots of all the trafodion tables
and exports them using the ExportSnapshot MapReduce job to a HDFS
location. The restore operations also use the ExportSnapshot
MapReduce job to import the snapshots back from the HDFS location
and the restore the tables.
The backup and restore were tested on clusters with Cloudera and
Horton Works distributions. it was also tested on the development
workstations

The Blueprint can be found at:
https://blueprints.launchpad.net/trafodion/+spec/full-backup-restore

Change-Id: I120a34eb4eb94e286577b4e6dc529ca528f0b846