Fix for bug 1329361 and minor improvements in hive scan
This is a joint checkin with Sandhya Sundaresan.

PreOpens during hive scan were not working correctly, cause the same range
to be read twice. The read triggered by the preopen was not used since it had
incorrect cursor name. This is now fixed.
Reduced number of LOB threads to 2
Removed multi cursor code since it is not used
Skip reading the first range if it has 0 bytes to be read.

Change-Id: I91cff41134490435165da7d59c955c7215b3c6b8