Fix to take care of internal statements leak
Internal statements in Trafodion engine use ExeCliInterface. When
this interface is instantiated in a scope of method, statements
and other related objects can be leaked.

Added a destructor in ExeCliInterface so that these objects
are destroyed correctly when the interface goes out of scope.

This fix possibly takes care of memory growth when many statements
are compiled.

Change-Id: Ic47c912ae153e9f83cca298b66da40bf214ef3e7