Errors reported for Update Stats on Hive tables
Attempts to execute Update Statistics statements on Hive
tables fail with error 1002 (catalog does not exist). The
catalog used to store Hive stats was no longer being
successfully created on demand. The Trafodion catalog is
now used instead, with a schema for Hive stats created
in it on first use. The tables for histograms and
histogram intervals for Hive tables are created in that
schema on first use.

Change-Id: Ib57d0af4a3da6f52f0544d6c2fce3e77d2e823c1
Closes-Bug: #1320397