Reworked the fix for bug 1418757
Interrupting the thread to cancel the pending IO was causing the
client to retry the connection and pause when the HTable was used
next time.

This fix allows the pending IO to complete normally. If it can't
complete normally within 30 seconds, the thread is interrupted
and the table is removed from cache.

Change-Id: I404ec7a91952301ef2bcb1a8d6448f8534a421d8