UDF, SHOWDDL, and DROP SCHEMA CASCADE
Support for UDFs was added early in Trafodion, but a few DDL
and security problems still existed:

o DROP FUNCTION RESTRICT does not fail when the UDF
  is referenced by a view
o DROP FUNCTION CASCADE does not drop referencing views
o References to undefined UDFs resulted in 4 error messages,
  one of which referenced the NEO catalog
o REVOKE and REVOKE ROLE do not check for dependent views
  when revoking EXECUTE from a UDF

All of these are addresses in this delivery except for
REVOKE of the EXECUTE privilege from a UDF that is
referenced by a view.

There was also a problem with indirect references to UDFs
not checking security, but that was corrected as part of
change 1177 that this change is built upon.

In addition, a recent change to prohibit unauthorized access
to metadata resulted in all SHOWDDL commands no longer working
for non DB__ROOT users.

DROP SCHEMA CASCADE did not drop sequence generators
not associated with a table.

Change-Id: I419334f5f33625f6aa0a0bcb70214ac181ba5cdf