Performance enhancement to avoid HBase file checks
Using hbaseAdmin to check if table exists in Hbase is an expensive call.

When a new compiler context is started, a call is made to see if all the
tables needed for privilege checking is performed.  There are currently
5 tables in privmgr metadata, so the hbaseAdmin is called five times.

A change has been made to just check for one privmgr metadata table.

Change-Id: If753f400fb3c51e18bcd2b3ed0edd26aaf300b21