Fix for updated tpcds_kit file.
Steve pointed out that when we download the latest TPCDS zip file with
the tools to build data files, the directory structure has
changed. Our script can now handle both versions of the file. Also
added a check for the existence of libhdfs.so.  This could be missing
if we download the hadoop tar file from the Internet instead of using
one that was generated by another Trafodion developer, with a
pre-built libhdfs.so.

Change-Id: Ib473ebc902063f05eee37284d208fe316424aea6