Address LP 1404951 - seabase/TEST010 failing
This commit changes the default value for COMP_BOOL_184 to 'ON'.
The effect of this is that MDAM probes will now fetch only one row
of data rather than the previous default of 100 rows.

NOTE: Because the original problem reported in LP 1404951 does not
seem to be reproducible any more, this check-in is being made to
improve the performance of MDAM problems.

HOWEVER, it is possible that this change will cause the original
problem to occur again.  Namely that an incorrect number of rows
could be returned for certain SQL queries (as originally reported
in LP 1404951.)  If anyone does run into this problem again, the
default value for COMP_BOOL_184 will need to be changed back to
'OFF' until the root cause for the incorrect number of rows issue
can be determined.

Change-Id: I95e5ae885bcb9122eb4e38d3b1f78c32aff5ef13