Local Hadoop JDBC test script and test_env short-circuit
For developers using local_hadoop environment, a new "swjdbc" script
is provided to run DCS jdbc tests, which is also a check/gate test.

The install_hadoop_regr_test_env also now checks if the hive tables
already exist. Since the tables are not modified by test runs, it
only needs to be run once per test system. This allows the jenkins
automation to call the script for every hive test run, but save several
minutes if it has previously been run.

Putting the short-circuit logic here allows it to change if additional
tables or data are added in future, without changing test automation.
Just in case, a force option is added to skip the check.

Change-Id: I96fd14957a1cf32dcd9c73127ccbb02472d51d4c