Fix for thread/memory leak on table create/drop
Objects that were put into our shared map between
coprocessors were not being removed.  This change
adds a remove of the objects on a postClose() of
the region.  A local thread was not being stopped
which lead to the TrxRegionEndpoint,TrxRegionObserver
and the thread itself not being GC'd as there was
still a reference.

Change-Id: Idaf3788d4a0b2065b93ec8746e3ea814615180b0