Metrics for batch RegionServer work
Changes are disabled by default.

The metrics will give statistics on the expected
benefit of batching our region coprocessor calls
to sending on a per-RegionServer basis. Values
measured are total region calls and total expected
RegionServer calls.  Also a list of the different
RS to Region ratios and the count of that ratio.

Example of what will be logged:
---------------------- BatchRS metrics
Number of total Region calls: 853
Number of total RegionServer calls: 237
---- Total number of calls by ratio
 2 RS / 3 Regions: 1
 3 RS / 13 Regions: 16
 2 RS / 2 Regions: 2
 3 RS / 6 Regions: 3
 3 RS / 14 Regions: 13
 3 RS / 15 Regions: 2
 3 RS / 12 Regions: 6
 1 RS / 1 Regions: 15
 2 RS / 4 Regions: 1
 3 RS / 16 Regions: 3
 2 RS / 6 Regions: 16
 1 RS / 2 Regions: 8
 3 RS / 11 Regions: 14
 3 RS / 3 Regions: 1
 --------------------------------------

Change-Id: Ibe45b16c763c2dae73a44fe1069b54945edeacaf