DROP SCHEMA CASCADE and other fixes
Corrects a problem (Launchpad bug 1435967 and 1437474) with DROP
SCHEMA CASCADE when a view is defined that references a UDF in
the schema.

Corrects a problem where DB__ROOT could not grant privileges on
objects owned by DB__ROOT except for tables and views.

Corrects a problem where histogram tables in shared schemas were
owned by the first user to run update statistics; the tables
are now owned by the schema owner regardless of which user runs
the UPDATE STATISTICS command first.

Change-Id: I3289d456ac0d1fdd3d7923892395ddf9229e47bf