Authentication internal error
Launchpad bug 1438856 reports an internal error when logging
on to Trafodion.  The problem is due to metadata no longer
selectable by all users.  When an MXOSRVR is returned to the
pool, the context is the last user to logon successfully. If
this is not DB__ROOT (or a user with SELECT privilege on the
AUTHS table), authentication fails with a privilege error,
which is translated to internal error for the user.

The fix is to enable internal query during authentication
to allow the select from the AUTHS table to succeed.

Change-Id: I7a02cb6ae6742e65d99771913843d949126dd948