Bug fix 1429923 and add error message in uninstall
Bug fix 1429923: traf_config_setup will now check if epel path entered
exists and if the file is an rpm. This questions will also only be asked
if this is the first time installing trafodion. Will warn users that if a
file is not entered that we will try to download it later. In traf_setup all nodes
are checked to see if epel is installed, if not and a path to the file
is not given then it will be downloaded and installed on all nodes
missing this rpm.

Added error message to trafodion_uninstaller if config file is not
found.

Change-Id: Ia4f254e238d2f4f6ca4fd2826c8a7a71797d9855