Drop view that references a sequence generator
Fixes a problem where a view that referenced a sequence generator
could not be dropped.  Code handling objects referenced by views
was not prepared for sequences.  Launchpad bug 1439319.

There are many other problems related to views referencing
sequences (e.g. see 1439316); this change only
corrects the problem of not being able to drop the view.

Change-Id: I8fbb8cbf565f9e6eef649152a0cec8e85d73afa4