Fix LP 1328250 - SQL Compiler hits "assert(FALSE)" in getIntConstValue
The SQL Compiler was abending and producing a core file in
getIntConstValue() when the Type of the integer constant was
PCIT::MBIGS.

This is code that is invoked only by the Native Expressions feature.
Currently, if the Native Expressions code encounters any PCODE
instruction that is designed for MBIGS or MBIGU type data, no native
expression is generated.  Instead, we continue running with the PCODE
instructions as they are.

However, if PCODE optimization converts an MBIG[SU] type instruction
into a simpler instruction such as a 64-bit move instruction, then
the Native Expressions code goes ahead and tries to produce a native
expression.  The problem is that, although PCODE optimization changes
to a simpler PCODE instruction, it does not change the data type of
any integer constant referenced by the instruction.  The constant is
left as being an MBIG[SU] type even though the instruction's operands
say that the operand is a standard integer type.

The fix is to use the operand's idea of the data type rather than
the constant's idea of the data type whenever the constant is of
MBIGS or MBIGU type.  This technique is already used by the code
when a Floating-Point type constant is referenced by a PCODE
instruction that says the operand is a simpler integer type.

Change-Id: I0e1aa28a51b8f5311ac51d0da5f385dc93bcd4eb