Allow longer compile times in compGeneral/TEST042
In the compGeneral suite, TEST042 checks that sql compile times are fast enough
in the release flavor.  This change allows longer times on OpenStack and VMware
virtual machines.  Allowed times for physical machines are unchanged.

The specific "failure" which requires this change was run on a VMware
virtual machine.  The difference from the expected output was:

  1794c1794
  < SUCCESS (less than 1ms)
  ---
  > FAIL (at least 1ms) Compile Time 00:00:00.001713
  1824c1824
  < SUCCESS (less than 1ms)
  ---
  > FAIL (at least 1ms) Compile Time 00:00:00.001821

The largest acceptable compile time on a virtual machine is now 1.999 ms, which
in the above format is "00:00:00.001999".

For physical machines, the largest value is unchanged at 1.399 ms.

This change requires detecting that the test is executing on a virtual machine.
That is found by looking in /sys/class/dmi/id/sys_vendor for "OpenStack"
or "VMware".

Change-Id: Ibd294c7e85917af1ecba993b80103e58f50d5ec7