LP 1444044, dup col definitions are not detected
During a create table or create view, duplicate col
names are not being detected and an error is not being
returned. This results in an incorect table or view
being created. An error may or may not be returned
later depending on how that object is being used.

That has been fixed to detect dup cols and return an
error message during create time.

Fixing it now so users dont create incorrect objects
by mistake.

Change-Id: I7ad772adcb067159ab80a487dead8dffc62bb546