Fix LP bug 1446402 - LIKE patterns longer than 127 chars don't work well
If a fixed part of a LIKE pattern is longer than 127 characters, then
you get "matches" on column values that should not match. An example of
such a pattern would be:
'%ABC...Z0123...9ABC...Z0123...9ABC...Z0123...9abcdefghij%'
where the fixed part [the part between two % (or _) characters]
is 128 characters long.

The root cause of the problem was another place in PCODE logic
where a signed char was being used to hold a length value.
By using an unsigned char, we can go up to 255 chars in a fixed part
of a LIKE pattern.  If a fixed part is longer than 255, the SQL Compiler
should not be attempting to use PCODE for the LIKE predicate so things
should be fine.

Change-Id: I2ff8e00dedeb3145602f57eed7418ea7b3c17a77